==> content/01_journal/02_data_acquisition.qmd <==
---
title: "Data Acquisition"
author: "Julius von Sulecki"
---

# Data Acquisition

Data for analysation can be imported in different ways. One way is to read data from csv or excel sheets, another is to import it directly from databases.

```{r}
library(RSQLite)
con <- RSQLite::dbConnect(drv = SQLite(), dbname = "../../00_data/02_chinook/Chinook_Sqlite.sqlite") #connect to database

library(DBI) #to print the tables print(dbListTables(con))

library(dplyr) 
print(tbl(con, "Album")) #to examine table from database

album_tbl <- tbl(con, "Album") %>% collect() #pull data in local memory

dbDisconnect(con) #disconnect database
```

# API

```{r}
library(glue)
name <- "Fred"
glue('My name is {name}.')

library(httr)
resp <- GET("https://swapi.dev/api/people/1/")

# Wrapped into a function
sw_api <- function(path) {
  url <- modify_url(url = "https://swapi.dev", path = glue("/api{path}"))
  resp <- GET(url)
  stop_for_status(resp) # automatically throws an error if a request did not succeed
}

resp <- sw_api("/people/1")
print(rep)
print(rawToChar(resp$content))

library(jsonlite)
print(fromJSON(rawToChar(resp$content)))
```

A .Renviron file was created in the home folder of the project, containing the token, password and userid. It can be accessed via `Sys.getenv('TOKEN')`

```{r}
print(Sys.getenv('TOKEN'))
response <- GET(glue("https://www.alphavantage.co/query?function=GLOBAL_QUOTE&symbol=WDI.DE&apikey={Sys.getenv('TOKEN')}"))
print(response)
```

# Web Scraping

Web scraping is for getting data from the internet if there is no open API for usage.

rvest is the library which is going to be used. ##Example:

```{r}
library(rvest)
library(stringr)

# get the URL for the wikipedia page with all S&P 500 symbols
url <- "https://en.wikipedia.org/wiki/List_of_S%26P_500_companies"
# use that URL to scrape the S&P 500 table using rvest
library(rvest)
sp_500 <- url %>%
  # read the HTML from the webpage
  read_html() %>%
  # Get the nodes with the id
  html_nodes(css = "#constituents") %>%
  # html_nodes(xpath = "//*[@id='constituents']"") %>% 
  # Extract the table and turn the list into a tibble
  html_table() %>% 
  .[[1]] %>% 
  as_tibble()
print(sp_500)
```

## And another example with the imdb movie database

```{r}
url  <- "https://www.imdb.com/chart/top/?ref_=nv_mv_250"
html <- url %>% 
  read_html()

rank <-  html %>% 
  html_nodes(css = ".titleColumn") %>% 
  html_text() %>% 
  # Extrag all digits between " " and ".\n" The "\" have to be escaped
  # You can use Look ahead "<=" and Look behind "?=" for this
  stringr::str_extract("(?<= )[0-9]*(?=\\.\\n)")%>% 
  # Make all values numeric
  as.numeric()
title <- html %>% 
  html_nodes(".titleColumn > a") %>% 
  html_text()
year <- html %>% 
  html_nodes(".titleColumn .secondaryInfo") %>%
  html_text() %>% 
  # Extract numbers
  stringr::str_extract(pattern = "[0-9]+") %>% 
  as.numeric()
people <- html %>% 
  html_nodes(".titleColumn > a") %>% 
  html_attr("title")
rating <- html %>% 
  html_nodes(css = ".imdbRating > strong") %>% 
  html_text() %>% 
  as.numeric()
num_ratings <- html %>% 
  html_nodes(css = ".imdbRating > strong") %>% 
  html_attr('title') %>% 
  # Extract the numbers and remove the comma to make it numeric values
  stringr::str_extract("(?<=based on ).*(?=\ user ratings)" ) %>% 
  stringr::str_replace_all(pattern = ",", replacement = "") %>% 
  as.numeric()
imdb_tbl <- tibble(rank, title, year, people, rating, num_ratings)
print(imdb_tbl)
```

## Functional Programming

```         
for (variable in vector) {
    
}
# Example: For Loop
numbers <- c(1:5)
for (i in numbers) {
   print(i)
}
```

would be written as

```         
numbers_list <- map(numbers, print)
```

### Example

```{r}
bike_data_lst <- fromJSON("../../00_data/bike_data.json")
bike_data_lst %>%
  purrr::pluck("productDetail", "variationAttributes", "values", 1, "displayValue") %>%
  print()
```

#Business Case not working.

# Challenge

## 1

The weather for a city in the north of Germany was collected for the past 7 days, including the sunrise, sunset, weathercode and the rain.

```{r}
resp <- GET(url = "https://api.open-meteo.com/v1/forecast?latitude=53.6751500&longitude=10.2259300&daily=weathercode,sunrise,sunset,rain_sum&past_days=7&forecast_days=1&timezone=Europe%2FLondon",
            add_headers(('accept'= 'application/json')))
html <- content(resp)
print(html)

weather_tbl <- tibble(html[["daily"]][["time"]], html[["daily"]][["weathercode"]], html[["daily"]][["sunrise"]], html[["daily"]][["sunset"]], html[["daily"]][["rain_sum"]])
weather_tbl %>% data.frame %>% print()
```

## 2

We scraped the electric mountainbike overview from radon bikes with the model names and the prices.

```{r}
# Challenge 2
library(tidyverse)
library(rvest)
library(xopen)
library(jsonlite)
library(glue)
library(stringi)


url <- "https://www.radon-bikes.de/e-bike/mountainbike/"

# People that worked on "The Dark Knight"
bike_names <- url %>% 
  read_html() %>% 
  html_nodes(".bikeTitle > h4") %>%
  html_text() 

bike_prices <- url %>% 
  read_html() %>% 
  html_nodes(".info > div > div > span") %>%
  html_text()
bike_prices <- bike_prices[seq(1,length(bike_prices),2)]

bikes_tbl <- tibble(bike_names, bike_prices)
bikes_tbl %>% data.frame %>% print()
```
